kaiax/supply: Move checkpoint.go to init.go
